Refactor: move over all peer related code from daemon
